backend/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
psycopg2-binary>=2.9.0
gunicorn>=21.0.0
python-decouple>=3.8
django-cors-headers>=4.0.0
channels>=4.0.0
channels-redis>=4.2.0
daphne>=4.0.0
django-filter>=23.0

# Development dependencies
coverage>=7.0.0
flake8>=6.0.0
flake8-docstrings>=1.6.0

